app/requirements/base.txt
fastapi==0.79.0
uvicorn[standard]==0.17.6
ujson==5.4.0
python-dotenv==0.20.0
sqlmodel==0.0.6
asyncpg==0.26.0
alembic==1.8.1
fastapi-pagination
wikipedia==1.4.0
numpy==1.23.4
pandas==1.5.1
pyforbes==1.0.3
google-auth==2.14.1
oauthlib==3.2.2
requests-oauthlib==1.3.1
python-decouple==3.6
requests==2.28.1
tweepy==4.12.1
urllib3==1.26.13
openai==0.25.0
sparqlwrapper==2.0.0
httpx==0.23.0
aiohttp==3.8.3
elastic-apm==6.13.2
